+++ intelmq/lib/message.py
@@ -9,7 +9,7 @@ import json
 import re
 import warnings
 from collections import defaultdict
-from typing import Any, Dict, Optional, Sequence, Union
+from typing import Any, Dict, Iterable, Optional, Sequence, Union
 
 import intelmq.lib.exceptions as exceptions
 import intelmq.lib.harmonization
@@ -374,7 +374,7 @@ class Message(dict):
     def __hash__(self):
         return int(self.hash(), 16)
 
-    def hash(self, *, filter_keys=frozenset(), filter_type="blacklist"):
+    def hash(self, *, filter_keys: Iterable = frozenset(), filter_type: str = "blacklist"):
         """Return a SHA256 hash of the message as a hexadecimal string.
         The hash is computed over almost all key/value pairs. Depending on
         filter_type parameter (blacklist or whitelist), the keys defined in
@@ -398,6 +398,7 @@ class Message(dict):
                 continue
 
             if filter_type == "whitelist" and key not in filter_keys:
+                print('not in whitelist')
                 continue
 
             if filter_type == "blacklist" and key in filter_keys:
@@ -407,6 +408,9 @@ class Message(dict):
             event_hash.update(b"\xc0")
             event_hash.update(utils.encode(repr(value)))
             event_hash.update(b"\xc0")
+            print('use', key, value, '->', event_hash.hexdigest())
+
+        print('result', event_hash.hexdigest())
 
         return event_hash.hexdigest()